requirements.txt
requests==2.31.0
pytest==7.4.3
pytest-html==4.1.1
pytest-json-report==1.5.0
python-dotenv==1.0.0
pyyaml==6.0.1
gitpython==3.1.40
docker==6.1.3
allure-pytest==2.13.2
parameterized==0.9.0